Inline all functional components
